requirements.txt
# OWASP Cheat Sheet Viewer Dependencies
# ======================================

# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0

# HTTP Client
httpx>=0.25.0

# HTML/Content Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Caching & Storage
diskcache>=5.6.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Utilities
tenacity>=8.2.0
python-dateutil>=2.8.0

# CrewAI (for checksheet crew)
crewai[tools]==1.3.0
